Downloads/content_monitor/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
python-dateutil>=2.8